Remove from peer.ID in fetcher
